--- home/models/learderboard.py
@@ -1,49 +1,0 @@
-from django.db import models
-
-
-
-# Event model
-class Leaderboard(models.Model):
-    """
-   
-    """
-
-    date = models.DateField(
-        help_text="The specific date for which the steps are recorded"
-    )
-    steps = models.IntegerField(help_text="Number of steps recorded")
-   
-    device = models.ForeignKey(
-        "Device",
-        on_delete=models.CASCADE,
-        help_text="Device the data is coming from",
-    )
-    account = models.ForeignKey(
-        "Account",
-        on_delete=models.CASCADE,
-        help_text="Account the data is linked to",
-    )
-    created = models.DateTimeField(
-        auto_now_add=True, help_text="Record creation timestamp"
-    )
-    updated = models.DateTimeField(
-        auto_now=True, help_text="Record updation timestamp"
-    )
-
-    @property
-    def distance_in_miles(self):
-        return m_to_mi(self.distance)
-
-    # Auto populate the account field from the device field
-    def save(self, *args, **kwargs):
-        self.account = self.device.account
-        super().save(*args, **kwargs)
-
-    def __str__(self):
-        return f"{self.account.email} | {self.date}"
-
-    class Meta:
-        ordering = ("-date",)
-        constraints = [
-            models.UniqueConstraint(
-                fields=["account", "date"], name="account_date"
